@@ -8,7 +8,7 @@
     "type-check": "tsc --noEmit",
     "type-check:watch": "npm run type-check -- --watch",
     "build": "webpack --mode=production --config webpack.config.js",
-    "start": "webpack-dev-server --config webpack.config.js --progress --colors --open --mode=development"
+    "start": "webpack-dev-server --config webpack.config.js --progress --open --mode=development"
   },
   "repository": {
     "type": "git",
@@ -28,29 +28,29 @@
     "@babel/preset-env": "^7.2.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.1.0",
-    "@types/react": "16.8.1",
-    "@types/react-dom": "16.0.11",
-    "@types/react-router-dom": "^4.3.4",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "@types/react-router-dom": "^5.3.3",
     "ajv": "^6.10.2",
     "babel-loader": "^8.0.5",
     "bootstrap": "^4.3.1",
-    "css-loader": "^2.1.1",
+    "css-loader": "^6.8.1",
     "datatables.net-dt": "^1.10.21",
-    "file-loader": "^3.0.1",
-    "fork-ts-checker-webpack-plugin": "0.5.2",
-    "html-webpack-plugin": "3.2.0",
+    "file-loader": "^6.2.0",
+    "fork-ts-checker-webpack-plugin": "^9.0.2",
+    "html-webpack-plugin": "^5.5.4",
     "jquery": "^3.5.1",
     "moment": "^2.24.0",
-    "node-sass": "^4.12.0",
-    "react": "16.8.4",
-    "react-dom": "16.8.4",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.21.0",
+    "react-router-dom": "^6.21.0",
+    "sass": "^1.69.5",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
     "typescript": "^3.2.2",
-    "webpack": "4.29.0",
-    "webpack-cli": "3.2.1",
-    "webpack-dev-server": "3.1.14"
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
